cloud/src/requirements.txt
jsonschema==4.19.0
functions-framework==3.*
google-auth-oauthlib==0.8.0
Flask==2.3.3
google-ads==25.1.0
google-api-core==2.24.0
firebase-functions==0.4.2